fix logger, reuse fs, move state package out of render
